requirements.txt
streamlit_agraph
streamlit == 1.13
pandas
Pillow